agent/requirements.txt
fastapi
uvicorn
playwright
python-dotenv
google-genai
websockets
playwright-stealth
